--- backend/config/APP_ARGO_SEARCH_RESULT_LIMIT
@@ -1,1 +1,0 @@
-1000
