--- backend/config/APP_OCP_SEARCH_RESULT_LIMIT
@@ -1,1 +1,0 @@
-1000
