+++ backend/config/APP_SEARCH_INTERVAL
@@ -1,0 +1,1 @@
+30000
